+++ dimension/dimension.go
@@ -2,6 +2,6 @@ package dimension
 
 //Dimension ...
 type Dimension struct {
-	Height uint16
-	Width  uint16
+	Height int
+	Width  int
 }
